refactor: simplify queue history list coverage
